% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anti_purl.R
\name{anti_purl}
\alias{anti_purl}
\title{Extracts contents of an R Markdown or Quarto file}
\usage{
anti_purl(file_path, clip = FALSE)
}
\arguments{
\item{file_path}{The path to the file.}

\item{clip}{Logical value indicating whether to copy the prose to the clipboard. Default is FALSE.}
}
\value{
If clip is FALSE, the function returns the extracted prose. If clip is TRUE, the function copies the prose to the clipboard.
}
\description{
This function extracts the prose from a file and optionally copies it to the clipboard. In this way, it is the opposite of the purl function.
}
\examples{
\dontrun{
anti_purl("/path/to/file.Rmd", clip = TRUE)
}
}
