% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_udpipe_model_path.R
\name{get_udpipe_model_path}
\alias{get_udpipe_model_path}
\title{Get the path to a language-specific `udpipe` model}
\usage{
get_udpipe_model_path(model_lang)
}
\arguments{
\item{model_name}{The name of the language model to identify}
}
\description{
This function will search for the desired udpipe language model path and
 download it if it is not found.
}
\examples{
\dontrun{
# Search and return the path to the Spanish language model
spanish_model <- get_udpipe_model_path(model_lang = "spanish")
}
get_udpipe_model_path()
}
\keyword{cleanNLP,}
\keyword{udpipe}
